% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2PATH.R
\name{add2PATH}
\alias{add2PATH}
\title{Add to PATH}
\usage{
add2PATH(dirs)
}
\arguments{
\item{dirs}{Character or character vector of directories to be added to the PATH variable.}
}
\description{
This function adds a directory to the system PATH variable
}
\examples{
add2PATH("~/src/mothur")
}
\seealso{
\code{\link{Sys.getenv}}, \code{\link{Sys.setenv}}
}
